WeakHold pattern with directional drift + volatility bursts
Adds weakHold to morning/close session: 9x volume like a real hold but with reduced volatility (0.3x) and a directional target. generateWeakHold mirrors generateHold's tight/loose chunk alternation, giving the WeakHoldDrift / WeakHoldFakeout label split. generateDrift gains a totalVolume parameter so chunks share a single target trajectory: progress = min(volumeLimit, volumeConsumed) / totalVolume, and cont threads currentTarget rather than endTarget so consecutive chunks pick up where the previous left off. Single-chunk callers pass volumeLimit=totalVolume and behave exactly as before. New label classes 11/12/13 (WeakHoldUp/Down/Fakeout) registered in labelToInt with magenta hues in btc_hold_pred chart. sim_volume now accepts a float bar size.

+++ scripts/visualization/sim_volume.py
@@ -242,7 +242,7 @@ def plot_volume_bars(bars, output_html, input_csv):
 
 if __name__ == '__main__':
     input_csv = sys.argv[1] if len(sys.argv) > 1 else 'data/test_hmm.csv'
-    volume_per_bar = int(sys.argv[2]) if len(sys.argv) > 2 else 10000
+    volume_per_bar = float(sys.argv[2]) if len(sys.argv) > 2 else 10000.0
     if len(sys.argv) > 3:
         output_html = sys.argv[3]
     else:
